api-portal/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
jinja2>=3.1.2
aiosqlite>=0.19.0
databases>=0.8.0
sqlalchemy>=2.0.0
httpx>=0.25.0
redis>=5.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
